backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
aiofiles==0.7.0
python-magic==0.4.24
requests==2.26.0
bcrypt==3.2.0
python-dateutil==2.8.2
email-validator==1.1.3
jinja2==3.0.1
itsdangerous==2.0.1
python-dotenv==0.19.0
cryptography==3.4.7
PyJWT==2.3.0
